pyproject.toml
[tool.poetry]
name = "hw-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1"
djangorestframework = "^3.15.2"
psycopg = "^3.2.1"
drf-yasg = "^1.21.7"
django-rest-swagger = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
